requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.0
billiard==3.6.4.0
binance==0.3
binance-connector==1.17.0
boto3==1.24.66
botocore==1.27.66
braintree==4.16.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coverage==6.4.4
cryptography==37.0.4
cycler==0.11.0
dateparser==1.1.1
Django==4.1.1
django-cors-headers==3.13.0
django-payments==1.0.0
django-phonenumber-field==6.4.0
django-redis==5.2.0
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
fastapi==0.85.0
fonttools==4.37.1
frozenlist==1.3.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==1.0.1
kiwisolver==1.4.4
kombu==5.2.4
matplotlib==3.5.3
multidict==6.0.2
numpy==1.23.2
package-name==0.1
packaging==21.3
pandas==1.4.4
phonenumberslite==8.12.54
Pillow==9.2.0
prompt-toolkit==3.0.33
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.5
regex==2022.3.2
requests==2.28.1
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
starlette==0.20.4
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
ujson==5.4.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
websockets==10.3
yarl==1.8.1
zope.interface==5.4.0
